+++ webengineviewer/src/webenginescript.h
@@ -36,7 +36,7 @@ WEBENGINEVIEWER_EXPORT QString setStyleToElement(const QString &elementStr, cons
 WEBENGINEVIEWER_EXPORT QString scrollDown(int pixel);
 WEBENGINEVIEWER_EXPORT QString scrollUp(int pixel);
 WEBENGINEVIEWER_EXPORT QString scrollPercentage(int percent);
-WEBENGINEVIEWER_EXPORT QString scrollToRelativePosition(int pos);
+WEBENGINEVIEWER_EXPORT QString scrollToRelativePosition(qreal pos);
 WEBENGINEVIEWER_EXPORT QString removeStyleToElement(const QString &element);
 WEBENGINEVIEWER_EXPORT QString clearSelection();
 }
